Created module for common functions
