+++ SerialDecrypt.py
@@ -8,25 +8,7 @@ import hashlib              # For SHA256
 import CONFIG               # Debug flags and constants
 import CoreFunctions as cf  # Common functions
 
-#PyCUDA Import
-import pycuda.driver as cuda
-import pycuda.autoinit
-from pycuda.compiler import SourceModule
-
 os.chdir(CONFIG.PATH)
-
-mod = SourceModule("""
-    #include <stdint.h>
-    __global__ void ArCatMap(uint8_t *in, uint8_t *out)
-    {
-        int nx = (2*blockIdx.x + blockIdx.y) % gridDim.x;
-        int ny = (blockIdx.x + blockIdx.y) % gridDim.y;
-        int blocksize = blockDim.x * blockDim.y * blockDim.z;
-        int InDex = ((gridDim.x)*blockIdx.y + blockIdx.x) * blocksize  + threadIdx.x;
-        int OutDex = ((gridDim.x)*ny + nx) * blocksize + threadIdx.x;
-        out[OutDex] = in[InDex];
-    }
-  """)
 
 # Driver function
 def Decrypt():
@@ -57,29 +39,20 @@ def Decrypt():
         cf.CatmapClear()
     
     timer[2] = time.time()
-    # Ar Phase: Cat-map Iterations
+    # Inverse Ar Phase: Cat-map Iterations
     cv2.imwrite("8output.png", imgAr)
-    dim = imgAr.shape
-    imgAr_In = np.asarray(imgAr).reshape(-1)
-    gpuimgIn = cuda.mem_alloc(imgAr_In.nbytes)
-    gpuimgOut = cuda.mem_alloc(imgAr_In.nbytes)
-
     while (cf.sha2Hash("8output.png")!=srchash):
-        cuda.memcpy_htod(gpuimgIn, imgAr_In)
-        func = mod.get_function("ArCatMap")
-        func(gpuimgIn, gpuimgOut, grid=(dim[0],dim[1],1), block=(dim[2],1,1))
-        cuda.memcpy_dtoh(imgAr_In, gpuimgOut)
-        imgAr = (np.reshape(imgAr_In,dim)).astype(np.uint8)
+        imgAr = cf.ArCatMap(imgAr)
         cv2.imwrite("8output.png", imgAr)
     timer[2] = time.time() - timer[2]
-    cv2.imwrite("8output.png", imgAr)
     overall_time = time.time() - overall_time
-
+    
     # Print timing statistics
-    print("Fractal XOR Inversion completed in " + str(timer[0]) +"s")
-    print("MT Unshuffle completed in " + str(timer[1]) +"s")
-    print("Arnold UnMapping completed in " + str(timer[2]) +"s")
-    print("Decryption took " + str(np.sum(timer)) + "s out of " + str(overall_time) + "s of net execution time")
+    if CONFIG.DEBUG_TIMER:
+        print("Fractal XOR Inversion completed in " + str(timer[0]) +"s")
+        print("MT Unshuffle completed in " + str(timer[1]) +"s")
+        print("Arnold Mapping completed in " + str(timer[2]) +"s")
+        print("Decryption took " + str(np.sum(timer)) + "s out of " + str(overall_time) + "s of net execution time")
 
 Decrypt()
 cv2.waitKey(0)
